+++ src/classes/scripts.ts
@@ -254,6 +254,8 @@ export class Scripts {
     const queueKeys = queue.keys;
     const opts: WorkerOptions = <WorkerOptions>queue.opts;
 
+    const metricsKey = queue.toKey(`metrics:${target}`);
+
     const keys = [
       queueKeys.active,
       queueKeys[target],
@@ -263,15 +265,15 @@ export class Scripts {
       queueKeys.events,
       queueKeys.meta,
       queueKeys.stalled,
+      metricsKey,
     ];
 
-    const keepJobs = pack(
+    const keepJobs =
       typeof shouldRemove === 'object'
         ? shouldRemove
         : typeof shouldRemove === 'number'
         ? { count: shouldRemove }
-        : { count: shouldRemove ? 0 : -1 },
-    );
+        : { count: shouldRemove ? 0 : -1 };
 
     const args = [
       job.id,
@@ -279,17 +281,21 @@ export class Scripts {
       propVal,
       typeof val === 'undefined' ? 'null' : val,
       target,
-      keepJobs,
       JSON.stringify({ jobId: job.id, val: val }),
       !fetchNext || queue.closing || opts.limiter ? 0 : 1,
       queueKeys[''],
       token,
-      opts.lockDuration,
-      job.opts?.parent?.id,
-      job.opts?.parent?.queue,
-      job.parentKey,
-      job.opts.attempts,
-      job.attemptsMade,
+      pack({
+        keepJobs,
+        lockDuration: opts.lockDuration,
+        parent: job.opts?.parent,
+        parentKey: job.parentKey,
+        attempts: job.opts.attempts,
+        attemptsMade: job.attemptsMade,
+        maxMetricsSize: opts.metrics?.maxDataPoints
+          ? opts.metrics?.maxDataPoints
+          : '',
+      }),
     ];
 
     return keys.concat(args);