+++ src/commands/moveToActive-8.lua
@@ -55,21 +55,31 @@ if jobId then
   if(maxJobs) then
     local rateLimiterKey = KEYS[6];
 
+    local groupKey
     if(ARGV[8]) then
-      local groupKey = string.match(jobId, "[^:]+$")
-      if groupKey ~= nil then
+      groupKey = string.match(jobId, "[^:]+$")
+      if groupKey ~= jobId then
         rateLimiterKey = rateLimiterKey .. ":" .. groupKey
       end
     end
 
-    local jobCounter = tonumber(rcall("INCR", rateLimiterKey))
+    local jobCounter
+    
+    if groupKey ~= nil then
+      if rateLimiterKey ~= KEYS[6] then
+        jobCounter = tonumber(rcall("INCR", rateLimiterKey))
+      end
+    else
+      jobCounter = tonumber(rcall("INCR", rateLimiterKey))
+    end
+
     -- check if rate limit hit
-    if jobCounter > maxJobs then
+    if jobCounter ~= nil and jobCounter > maxJobs then
       local exceedingJobs = jobCounter - maxJobs
       local expireTime = tonumber(rcall("PTTL", rateLimiterKey))
       local delay = expireTime + ((exceedingJobs - 1) * ARGV[7]) / maxJobs;
       local timestamp = delay + tonumber(ARGV[4])
-      
+
       -- put job into delayed queue
       rcall("ZADD", KEYS[7], timestamp * 0x1000 + bit.band(jobCounter, 0xfff), jobId);
       rcall("XADD", KEYS[4], "*", "event", "delayed", "jobId", jobId, "delay", timestamp);
@@ -92,9 +102,7 @@ if jobId then
   -- get a lock
   rcall("SET", lockKey, ARGV[2], "PX", ARGV[3])
 
-  moveJobFromWaitToActive(KEYS[3], KEYS[4], jobKey, jobId, ARGV[4])
+  moveJobFromWaitToActive(KEYS[1], KEYS[3], KEYS[4], jobKey, jobId, ARGV[4])
 
   return {rcall("HGETALL", jobKey), jobId} -- get job data
-else
-  rcall("XADD", KEYS[4], "*", "event", "drained");
 end