+++ tests/test_flow.ts
@@ -357,7 +357,7 @@ describe('flows', () => {
       );
 
       const processingParent = new Promise<void>((resolve, reject) => [
-        (parentProcessor = async (job: Job) => {
+        (parentProcessor = async () => {
           try {
             resolve();
           } catch (err) {
@@ -432,11 +432,11 @@ describe('flows', () => {
 
   describe('when continually adding jobs', async () => {
     it('adds jobs that do not exists', async () => {
-      const worker = new Worker(queueName, async (job: Job) => {}, {
+      const worker = new Worker(queueName, async () => {}, {
         connection,
       });
 
-      const completing1 = new Promise<void>((resolve, reject) => {
+      const completing1 = new Promise<void>(resolve => {
         worker.on('completed', (job: Job) => {
           if (job.id === 'wed') {
             resolve();
@@ -481,7 +481,7 @@ describe('flows', () => {
 
       await completing1;
 
-      const completing2 = new Promise<void>((resolve, reject) => {
+      const completing2 = new Promise<void>(resolve => {
         worker.on('completed', (job: Job) => {
           if (job.id === 'thu') {
             resolve();
@@ -627,7 +627,7 @@ describe('flows', () => {
         parentProcessor,
         processedGrandChildren = 0,
         processedChildren = 0;
-      const processingChildren = new Promise<void>((resolve, reject) => {
+      const processingChildren = new Promise<void>(resolve => {
         childrenProcessor = async (job: Job) => {
           processedChildren++;
           await delay(20);
@@ -789,7 +789,7 @@ describe('flows', () => {
     const parentWorker = new Worker(parentQueueName, parentProcessor, {
       connection,
     });
-    const childrenWorker = new Worker(queueName, async job => {}, {
+    const childrenWorker = new Worker(queueName, async () => {}, {
       connection,
       limiter: {
         max: 1,